+++ src/extension.ts
@@ -6,7 +6,7 @@ import {
     newMacroCommand,
     openMacrosDir,
 } from "./macroStorage";
-import { runMacroCommand } from "./runMacroCommand";
+import { runMacroCommand, runMacroCommandWithFilePicker } from "./runMacroCommand";
 
 export var macrosUri: vscode.Uri | null = null;
 
@@ -17,8 +17,23 @@ export function activate(context: vscode.ExtensionContext) {
     macrosUri = vscode.Uri.joinPath(storagePath, "macros");
 
     context.subscriptions.push(
+        vscode.commands.registerCommand("MacroRunner.runMacro", async () => {
+            vscode.window.withProgress(
+                { location: vscode.ProgressLocation.Window, title: "macro running..." },
+                runMacroCommand
+            );
+        }),
+        vscode.commands.registerCommand("MacroRunner.runMacroLF", async () => {
+            vscode.window.withProgress(
+                {
+                    location: vscode.ProgressLocation.Window,
+                    title: "(Large File) macro running...",
+                },
+                runMacroCommandWithFilePicker
+            );
+        }),
+
         vscode.commands.registerCommand("MacroRunner.newMacro", newMacroCommand),
-        vscode.commands.registerCommand("MacroRunner.runMacro", runMacroCommand),
         vscode.commands.registerCommand("MacroRunner.loadMacro", loadMacroCommand),
         vscode.commands.registerCommand("MacroRunner.saveMacro", saveMacroCommand),
         vscode.commands.registerCommand("MacroRunner.removeMacro", removeMacroCommand),