+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
-## Upcoming features
-- Nothing for now
+## 1.1.0 2022/04/06
+- Fixed run macro command not working when you have an output log open
+- Fixed new macro opening in the wrong 'column'
+- Added a workaround for VS-Code extension host not being able to find large files. You will now have to manually find these the files
+- Added 1000 chars of verbose log to the stack trace for diagnostic purposes
+- Running a macro now shows a progress spinner at the bottom there
+- Added some more example use-cases to the repo
 
 ## 1.0.0 - 2022/03/28
 - Added all features